refactor: integrate developer cards into unified DeveloperCard
