+++ src/pages/main/main-page.tsx
@@ -10,11 +10,10 @@ import { getMyProjectsAllStatuses } from '@apis/projects';
 import { getReports } from '@apis/report/report-queries';
 import { useAuth } from '@clerk/clerk-react';
 import LoginRequiredCard from '@ui/LoginRequiredCard';
-import MainProjectCard from '@components/common/MainProjectCard';
-import RecommendDeveloperCard from '@components/common/RecommendDeveloperCard';
-import { RecommendDeveloperCardSkeletonList } from '@components/common/RecommendDeveloperCardSkeleton';
-import RecommendProjectCard from '@components/common/RecommendProjectCard';
-import { RecommendProjectCardSkeletonList } from '@components/common/RecommendProjectCardSkeleton';
+import { ProjectCard } from '@components/project/ProjectCard';
+import { DeveloperCard } from '@components/developer/DeveloperCard';
+import { ProjectCardSkeletonList } from '@components/project/ProjectCardSkeleton';
+import { RecommendDeveloperCardSkeletonList } from '@components/developer/RecommendDeveloperCardSkeleton';
 import ReportRequiredCard from '@ui/ReportRequiredCard';
 import Skeleton from '@ui/Skeleton';
 import {
@@ -564,7 +563,7 @@ const MainPage = () => {
                 </div>
               ))
             : highlightProjects.map((project) => (
-                <MainProjectCard
+                <ProjectCard variant="grid"
                   key={project.id}
                   categoryLabel={project.categoryLabel}
                   deadlineLabel={project.deadlineLabel}
@@ -612,12 +611,13 @@ const MainPage = () => {
                     </div>
                   ) : (
                     recommendedProfiles.map((profile) => (
-                      <RecommendDeveloperCard
+                      <DeveloperCard
+                        variant="recommend"
                         key={profile.id}
                         role={profile.role}
                         roleTone={profile.roleTone}
-                        nickname={profile.nickname}
-                        profileImageUrl={profile.profileImageUrl}
+                        nickname={profile.nickname ?? ''}
+                        profileImageUrl={profile.profileImageUrl ?? ''}
                         introduction={profile.introduction}
                         domains={profile.badges?.map((badge) => ({ label: badge.label }))}
                         techStack={profile.techStack}
@@ -661,7 +661,7 @@ const MainPage = () => {
                           ? projectBookmarkMap[targetId] != null
                           : (project.bookmarked ?? false);
                         return (
-                          <RecommendProjectCard
+                          <ProjectCard variant="recommend"
                             key={project.id}
                             categoryLabel={project.categoryLabel}
                             deadlineLabel={project.deadlineLabel}
