+++ src/pages/search/project/project-search-page.tsx
@@ -11,12 +11,11 @@ import Pagination from '@ui/Pagination';
 import ProjectFiltersBar, {
   PROJECT_FILTERS,
   type ProjectFilterKey,
-} from '@components/common/ProjectFilterBar';
-import ProjectLg from '@components/common/ProjectLg';
-import ProjectSm from '@components/common/ProjectSm';
-import RecommendProjectCardSkeleton, {
-  RecommendProjectCardSkeletonList,
-} from '@components/common/RecommendProjectCardSkeleton';
+} from '@components/project/ProjectFilterBar';
+import { ProjectCard } from '@components/project/ProjectCard';
+import ProjectCardSkeleton, {
+  ProjectCardSkeletonList,
+} from '@components/project/ProjectCardSkeleton';
 import { useInitialSkeletonGate } from '@hooks/useInitialSkeletonGate';
 import { useMyReportsExist } from '@hooks/useMyReportsExist';
 import { useProjectFilter } from '@hooks/useProjectFilters';
@@ -325,7 +324,7 @@ export default function ProjectSearchPage() {
         <div className="scrollbar-hide flex gap-6 overflow-x-auto py-2">
           {Array.from({ length: 4 }, (_, i) => (
             <div key={i} className="w-[260px] shrink-0">
-              <RecommendProjectCardSkeleton variant="compact" />
+              <ProjectCardSkeleton variant="compact" />
             </div>
           ))}
         </div>
@@ -340,7 +339,7 @@ export default function ProjectSearchPage() {
           {recommendedPreview.map((p) => {
             const ov = bookmarkOverrides[Number(p.id)];
             return (
-              <ProjectSm
+              <ProjectCard variant="compact"
                 key={p.id}
                 categoryLabel={p.categoryLabel}
                 deadlineLabel={p.deadlineLabel}
@@ -385,7 +384,7 @@ export default function ProjectSearchPage() {
       {/* 프로젝트 리스트 */}
       <div className="flex flex-col gap-6">
         {showProjectListSkeleton && (
-          <RecommendProjectCardSkeletonList count={3} className="py-2" />
+          <ProjectCardSkeletonList count={3} className="py-2" />
         )}
 
         {!showProjectListSkeleton && isError && (
@@ -400,7 +399,7 @@ export default function ProjectSearchPage() {
           !isError &&
           projects.length > 0 &&
           projects.map((p) => (
-            <ProjectLg
+            <ProjectCard variant="list"
               key={p.id}
               {...p}
               onClick={() => handleProjectClick(p.id)}
